リダイレクト方式の廃止
