+++ src/pages/index.tsx
@@ -1,43 +1,9 @@
-import Head from 'next/head';
-import { Button } from '@mui/material';
-import { useEffect, useState } from 'react';
-import Modal from '@/components/Modal';
-import { useRouter } from 'next/router';
+import TopTemplate from '@/components/templates/Top';
 
 export default function Home() {
-  const router = useRouter();
-  const [open, setOpen] = useState<boolean>(false);
-
-  const handleOpen = async () => {
-    setOpen(true);
-    await router.push('/', '/modal/sample', { shallow: true });
-  };
-  const handleClose = async () => {
-    setOpen(false);
-    await router.push('/', '/', { shallow: true });
-  };
-  useEffect(() => {
-    if (!!router.query.modalName) {
-      handleOpen();
-    }
-  }, []);
-
   return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main>
-        <h1>モーダル+url+レスポンシブサンプル</h1>
-        <Button onClick={handleOpen}>モーダルを開く</Button>
-        <Modal open={open} handleClose={handleClose}>
-          <h3>モーダルサンプル</h3>
-          これがサンプルです
-        </Modal>
-      </main>
-    </>
+    <main>
+      <TopTemplate modalOpen={false} />
+    </main>
   );
 }
